server: changed recipient show validation
